iio: light: Add BH1745 RGBc ambient sensor driver
The BH1745 light sensor is a device that can be connected via i2c
and used to measure color intensity. It captures data for each
color channel, red, blue, and green, as well as clarity, which
are then stored in four 16-bit registers.

To communicate with userspace, the driver relies on the iio
framework.

Signed-off-by: Corentin Teissier <corentinteiss@yahoo.fr>
Co-developed-by: Andi Shyti <andi.shyti@studenti.polito.it>
Signed-off-by: Andi Shyti <andi.shyti@studenti.polito.it>